move overview-only attrs to overview component
